refactor(api): split calendar routes into features/calendar, extract event serializer
